Refactor collage text editor to standalone component
