+++ src/components/collage/components/CanvasCollagePreview.js
@@ -1,179 +1,11 @@
 import React, { useRef, useEffect, useState, useCallback, useMemo } from 'react';
-import { Box, IconButton, Typography, TextField, Slider, FormControl, InputLabel, Select, MenuItem, Button, Tabs, Tab, Tooltip, useMediaQuery, ToggleButtonGroup, ToggleButton, Dialog, DialogTitle, DialogContent, DialogContentText, DialogActions } from "@mui/material";
+import { Box, IconButton, Typography, useMediaQuery } from "@mui/material";
 import { useTheme, styled, alpha } from "@mui/material/styles";
-import { Add, OpenWith, Check, Edit, FormatColorText, Close, FormatSize, BorderOuter, FormatBold, FormatItalic, FontDownload, ControlCamera, SwapHoriz, SwapVert, Colorize, ChevronLeft, ChevronRight, Palette, Brush, RotateLeft, Restore } from '@mui/icons-material';
+import { Add, OpenWith, Check } from '@mui/icons-material';
 import { layoutDefinitions } from '../config/layouts';
-import fonts from '../../../utils/fonts';
-
-// Color presets for text colors
-const TEXT_COLOR_PRESETS = [
-  { color: '#ffffff', name: 'White' },
-  { color: '#000000', name: 'Black' },
-  { color: '#ff0000', name: 'Red' },
-  { color: '#00ff00', name: 'Green' },
-  { color: '#0000ff', name: 'Blue' },
-  { color: '#ffff00', name: 'Yellow' },
-  { color: '#ff00ff', name: 'Magenta' },
-  { color: '#00ffff', name: 'Cyan' },
-];
-
-// Horizontal scrollable container for horizontal scrolling sections
-const HorizontalScroller = styled(Box)(({ theme }) => ({
-  display: 'flex',
-  overflowX: 'auto',
-  overflowY: 'hidden', // Explicitly prevent vertical scrolling
-  scrollbarWidth: 'none',  // Firefox
-  '&::-webkit-scrollbar': {
-    display: 'none',  // Chrome, Safari, Opera
-  },
-  msOverflowStyle: 'none',  // IE, Edge
-  gap: theme.spacing(1),
-  padding: theme.spacing(0.75, 0),
-  position: 'relative',
-  scrollBehavior: 'smooth',
-  alignItems: 'center',  // Center items vertically
-  justifyContent: 'flex-start',  // Start alignment for consistent scrolling
-  minHeight: 32,
-  maxWidth: '100%', // Ensure it doesn't exceed container width
-  width: '100%', // Take full width of parent
-  boxSizing: 'border-box', // Include padding in width calculation
-  // Contain content to prevent layout shift
-  contain: 'layout style',
-  // Smoother momentum scrolling (for Safari)
-  WebkitOverflowScrolling: 'touch',
-  // Show that content is scrollable on mobile
-  overscrollBehavior: 'contain',
-  // Prevent accidental touch interactions
-  touchAction: 'pan-x', // Allow horizontal panning only
-  // Ensure proper mobile touch scrolling
-  '@media (hover: none)': {
-    // Mobile-specific styles
-    overflowX: 'scroll', // Force scroll on mobile
-    '-webkit-overflow-scrolling': 'touch',
-    scrollSnapType: 'x mandatory',
-    scrollSnapAlign: 'start',
-  },
-}));
-
-// Improved ScrollButton for consistent appearance
-const ScrollButton = styled(IconButton)(({ theme, direction }) => ({
-  position: 'absolute',
-  top: '50%',
-  transform: 'translateY(-50%)',
-  zIndex: 10,
-  // Consistent styling across all devices
-  backgroundColor: theme.palette.mode === 'dark' 
-    ? alpha(theme.palette.background.paper, 0.9)
-    : alpha(theme.palette.background.paper, 0.95),
-  // Better shadow for depth without overwhelming the UI
-  boxShadow: `0 2px 6px ${theme.palette.mode === 'dark' 
-    ? 'rgba(0,0,0,0.3)' 
-    : 'rgba(0,0,0,0.15)'}`,
-  // Clean border
-  border: `1px solid ${theme.palette.mode === 'dark'
-    ? alpha(theme.palette.divider, 0.5)
-    : theme.palette.divider}`,
-  // Primary color for better visibility
-  color: theme.palette.primary.main,
-  '&:hover': {
-    backgroundColor: theme.palette.mode === 'dark' 
-      ? alpha(theme.palette.background.paper, 0.95)
-      : alpha(theme.palette.background.default, 0.98),
-    color: theme.palette.primary.dark,
-    transform: 'translateY(-50%) scale(1.05)',
-    boxShadow: `0 3px 8px ${theme.palette.mode === 'dark' 
-      ? 'rgba(0,0,0,0.4)' 
-      : 'rgba(0,0,0,0.2)'}`,
-  },
-  // Consistent positioning for both directions
-  ...(direction === 'left' ? { left: -4 } : { right: -4 }),
-  // Consistent sizing across devices
-  width: 24,
-  height: 24,
-  minWidth: 'unset',
-  padding: 0,
-  // Consistent circular shape on all devices
-  borderRadius: '50%',
-  // Better transition for hover states
-  transition: theme.transitions.create(
-    ['background-color', 'color', 'box-shadow', 'transform', 'opacity'], 
-    { duration: theme.transitions.duration.shorter }
-  ),
-}));
-
-// Improved ScrollIndicator with subtle gradient
-const ScrollIndicator = styled(Box, {
-  shouldForwardProp: (prop) => prop !== 'isVisible'
-})(({ theme, direction, isVisible }) => ({
-  position: 'absolute',
-  top: 0,
-  bottom: 0,
-  width: 30,
-  pointerEvents: 'none',
-  zIndex: 2,
-  opacity: isVisible ? 1 : 0,
-  transition: 'opacity 0.3s ease',
-  background: direction === 'left'
-    ? `linear-gradient(90deg, ${theme.palette.mode === 'dark' 
-        ? 'rgba(0,0,0,0.9)' 
-        : 'rgba(255,255,255,0.9)'} 0%, transparent 100%)`
-    : `linear-gradient(270deg, ${theme.palette.mode === 'dark' 
-        ? 'rgba(0,0,0,0.9)' 
-        : 'rgba(255,255,255,0.9)'} 0%, transparent 100%)`,
-  ...(direction === 'left' ? { left: 0 } : { right: 0 })
-}));
-
-// Create a color swatch component for text color selection
-const ColorSwatch = styled(Box)(({ theme, selected, size = 'medium' }) => {
-  const sizeMap = {
-    small: { width: 24, height: 24, borderWidth: selected ? 2 : 1 },
-    medium: { width: 36, height: 36, borderWidth: selected ? 3 : 2 },
-  };
-  const dimensions = sizeMap[size] || sizeMap.medium;
-  
-  return {
-    width: dimensions.width,
-    height: dimensions.height,
-    borderRadius: '50%',
-    cursor: 'pointer',
-    boxSizing: 'border-box',
-    border: selected ? `${dimensions.borderWidth}px solid ${theme.palette.primary.main}` : `${dimensions.borderWidth}px solid #ffffff`,
-    boxShadow: selected 
-      ? `0 0 0 ${Math.max(1, dimensions.borderWidth - 1)}px ${theme.palette.primary.main}` 
-      : '0 0 0 1px rgba(0,0,0,0.1)',
-    transition: theme.transitions.create(
-      ['transform', 'box-shadow'],
-      { duration: theme.transitions.duration.shorter }
-    ),
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    flexShrink: 0,
-    '&:hover': {
-      transform: size === 'small' ? 'scale(1.2)' : 'scale(1.15)',
-      boxShadow: '0 4px 10px rgba(0,0,0,0.15)'
-    },
-    '&:active': {
-      transform: 'scale(0.95)',
-    }
-  };
-});
+import CaptionEditor from './CaptionEditor';
 
 
-
-// Helper function to determine if a color is dark (for contrast)
-const isDarkColor = (hexColor) => {
-  // Convert hex to RGB
-  const r = parseInt(hexColor.substr(1, 2), 16);
-  const g = parseInt(hexColor.substr(3, 2), 16);
-  const b = parseInt(hexColor.substr(5, 2), 16);
-  
-  // Calculate brightness (YIQ formula)
-  const brightness = (r * 299 + g * 587 + b * 114) / 1000;
-  
-  // Return true if color is dark
-  return brightness < 128;
-};
 
 /**
  * Helper function to get border pixel size based on percentage and component width
@@ -486,13 +318,8 @@ const CanvasCollagePreview = ({
   const [isDragging, setIsDragging] = useState(false);
   const [dragStart, setDragStart] = useState({ x: 0, y: 0 });
   const [textEditingPanel, setTextEditingPanel] = useState(null);
-  const [activeTextSetting, setActiveTextSetting] = useState(null);
   const [touchStartDistance, setTouchStartDistance] = useState(null);
   const [touchStartScale, setTouchStartScale] = useState(1);
-  const [resetDialogOpen, setResetDialogOpen] = useState(false);
-  const [resetDialogData, setResetDialogData] = useState({ type: null, panelId: null, propertyName: null });
-  const [activeSlider, setActiveSlider] = useState(null); // Track which slider is being controlled
-  const textFieldRefs = useRef({});
   const lastInteractionTime = useRef(0);
   const hoverTimeoutRef = useRef(null);
   const touchStartInfo = useRef(null);
@@ -508,23 +335,6 @@ const CanvasCollagePreview = ({
     return componentWidth / BASE_CANVAS_WIDTH;
   }, [componentWidth]);
 
-  // State for text color scroll indicators
-  const [textColorLeftScroll, setTextColorLeftScroll] = useState(false);
-  const [textColorRightScroll, setTextColorRightScroll] = useState(false);
-  
-  // Refs for text color picker
-  const textColorScrollerRef = useRef(null);
-  const textColorPickerRef = useRef(null);
-  
-  // Saved custom text color state
-  const [savedCustomTextColor, setSavedCustomTextColor] = useState(() => {
-    const storedColor = localStorage.getItem('memeTextCustomColor');
-    return storedColor || '#ffffff';
-  });
-
-  // Check if there's a saved custom color that's different from preset colors
-  const hasSavedCustomTextColor = savedCustomTextColor && !TEXT_COLOR_PRESETS.some(c => c.color === savedCustomTextColor);
-
   // Get layout configuration
   const layoutConfig = useMemo(() => {
     return selectedTemplate ? createLayoutConfig(selectedTemplate, panelCount) : null;
@@ -533,67 +343,7 @@ const CanvasCollagePreview = ({
   // Calculate border pixels
   const borderPixels = getBorderPixelSize(borderThickness, componentWidth);
 
-  // Text color scroll handling functions - using the same pattern as CollageSettingsStep.js
-  const scrollTextColorLeft = () => {
-    if (textColorScrollerRef.current) {
-      // Calculate a consistent scroll distance based on container width
-      const scrollDistance = Math.min(textColorScrollerRef.current.clientWidth * 0.5, 200);
-      textColorScrollerRef.current.scrollBy({ left: -scrollDistance, behavior: 'smooth' });
-      
-      // Update scroll indicators after scrolling
-      setTimeout(() => {
-        handleTextColorScroll();
-      }, 350); // Slightly longer timeout to ensure scroll completes
-    }
-  };
 
-  const scrollTextColorRight = () => {
-    if (textColorScrollerRef.current) {
-      // Calculate a consistent scroll distance based on container width
-      const scrollDistance = Math.min(textColorScrollerRef.current.clientWidth * 0.5, 200);
-      textColorScrollerRef.current.scrollBy({ left: scrollDistance, behavior: 'smooth' });
-      
-      // Update scroll indicators after scrolling
-      setTimeout(() => {
-        handleTextColorScroll();
-      }, 350); // Slightly longer timeout to ensure scroll completes
-    }
-  };
-
-  const checkTextColorScrollPosition = useCallback(() => {
-    if (!textColorScrollerRef.current) return;
-    
-    const container = textColorScrollerRef.current;
-    const { scrollLeft, scrollWidth, clientWidth } = container;
-    
-    // Use a small threshold to detect scrollability
-    const threshold = 5;
-    
-    // Check if we can scroll left (not at the beginning)
-    const hasLeft = scrollLeft > threshold;
-    
-    // Check if we can scroll right (not at the end)
-    const hasRight = scrollLeft < scrollWidth - clientWidth - threshold;
-    
-    // Update state directly - the optimization was causing issues
-    setTextColorLeftScroll(hasLeft);
-    setTextColorRightScroll(hasRight);
-  }, []);
-
-  const handleTextColorScroll = useCallback(() => {
-    checkTextColorScrollPosition();
-  }, [checkTextColorScrollPosition]);
-
-  // Handle custom text color selection
-  const handleCustomTextColorChange = (e) => {
-    const newColor = e.target.value;
-    setSavedCustomTextColor(newColor);
-    localStorage.setItem('memeTextCustomColor', newColor);
-    // Apply to current panel if editing
-    if (textEditingPanel) {
-      handleTextChange(textEditingPanel, 'color', newColor);
-    }
-  };
 
   // Load images when they change
   useEffect(() => {
@@ -1283,11 +1033,8 @@ const CanvasCollagePreview = ({
     const isOpening = textEditingPanel !== panelId;
     setTextEditingPanel(textEditingPanel === panelId ? null : panelId);
     
-    // Always set to content tab when opening
+    // Auto-scroll to show the caption editor after it opens
     if (isOpening) {
-      setActiveTextSetting('content');
-      
-      // Auto-scroll to show the caption editor after it opens
       setTimeout(() => {
         const panel = panelRects.find(p => p.panelId === panelId);
         if (panel && containerRef.current) {
@@ -1430,265 +1177,18 @@ const CanvasCollagePreview = ({
 
   const handleTextClose = useCallback(() => {
     setTextEditingPanel(null);
-    setActiveTextSetting(null);
   }, []);
 
 
 
-  const handleTextChange = useCallback((panelId, property, value) => {
-    if (updatePanelText) {
-      const currentText = panelTexts[panelId] || {};
-      let updatedText = {
-        ...currentText,
-        [property]: value
-      };
-      
-      // Normalize font weight to numbers for better canvas compatibility
-      if (property === 'fontWeight') {
-        // Convert string weights to numbers for consistency
-        if (value === 'normal' || value === '400') {
-          updatedText[property] = 400;
-        } else if (value === 'bold' || value === '700') {
-          updatedText[property] = 700;
-        } else if (typeof value === 'string') {
-          // Convert other string numbers to actual numbers
-          const numValue = parseInt(value, 10);
-          updatedText[property] = Number.isNaN(numValue) ? 400 : numValue;
-        } else {
-          // Already a number
-          updatedText[property] = value;
-        }
-        
 
-      }
-      
-      // Set default font size only when first adding content to an empty text field
-      // This prevents font size from changing while the user is editing existing text
-      if (property === 'content' && value && value.trim()) {
-        const hadPreviousContent = currentText.content && currentText.content.trim();
-        const hasExplicitFontSize = currentText.fontSize !== undefined;
-        
-        // Only set default font size if there was no previous content AND no explicit font size set
-        if (!hadPreviousContent && !hasExplicitFontSize) {
-          updatedText = {
-            ...updatedText,
-            fontSize: lastUsedTextSettings.fontSize || 26 // Use default font size
-          };
-        }
-      }
-      
-      updatePanelText(panelId, updatedText);
-    }
-  }, [panelTexts, updatePanelText, panelRects, calculateOptimalFontSize]);
-
-  // Reset dialog handlers
-  const handleResetClick = useCallback((type, panelId, propertyName) => {
-    setResetDialogData({ type, panelId, propertyName });
-    setResetDialogOpen(true);
-  }, []);
-
-  const handleResetConfirm = useCallback(() => {
-    const { type, panelId, propertyName } = resetDialogData;
-    
-    if (propertyName === 'fontSize') {
-      // For font size, we need to clear the explicit fontSize and let the system recalculate
-      const currentText = panelTexts[panelId] || {};
-      const hasActualText = currentText.content && currentText.content.trim();
-      
-      if (hasActualText) {
-        // Clear explicit fontSize to let optimal size be recalculated
-        handleTextChange(panelId, propertyName, undefined);
-      } else {
-        // Use default size for empty text
-        handleTextChange(panelId, propertyName, lastUsedTextSettings.fontSize || 26);
-      }
-    } else {
-      // For all other properties, set to undefined to use default fallback logic
-      handleTextChange(panelId, propertyName, undefined);
-    }
-    
-    setResetDialogOpen(false);
-    setResetDialogData({ type: null, panelId: null, propertyName: null });
-  }, [resetDialogData, handleTextChange, panelTexts, lastUsedTextSettings]);
-
-  const handleResetCancel = useCallback(() => {
-    setResetDialogOpen(false);
-    setResetDialogData({ type: null, panelId: null, propertyName: null });
-  }, []);
-
-  // Helper function to check if a value matches its default
-  const isValueAtDefault = useCallback((panelId, propertyName) => {
-    const currentValue = panelTexts[panelId]?.[propertyName];
-    
-    if (propertyName === 'fontSize') {
-      // For fontSize, check if it's undefined (using calculated default) or matches lastUsedTextSettings
-      return currentValue === undefined || currentValue === (lastUsedTextSettings.fontSize || 26);
-    }
-    
-    if (propertyName === 'strokeWidth') {
-      // For strokeWidth, use the same logic as the rendering code
-      const defaultStrokeWidth = lastUsedTextSettings.strokeWidth || 2;
-      return currentValue === undefined || currentValue === defaultStrokeWidth;
-    }
-    
-    if (propertyName === 'textPositionX') {
-      const defaultPositionX = lastUsedTextSettings.textPositionX || 0;
-      return currentValue === undefined || currentValue === defaultPositionX;
-    }
-    
-    if (propertyName === 'textPositionY') {
-      const defaultPositionY = lastUsedTextSettings.textPositionY || 0;
-      return currentValue === undefined || currentValue === defaultPositionY;
-    }
-    
-    if (propertyName === 'textRotation') {
-      const defaultRotation = lastUsedTextSettings.textRotation || 0;
-      return currentValue === undefined || currentValue === defaultRotation;
-    }
-    
-    return false;
-  }, [panelTexts, lastUsedTextSettings]);
-
-  // Helper function to get current value for display
-  const getCurrentValue = useCallback((panelId, propertyName) => {
-    const currentValue = panelTexts[panelId]?.[propertyName];
-    
-    if (propertyName === 'fontSize') {
-      const panelText = panelTexts[panelId] || {};
-      const hasActualText = panelText.content && panelText.content.trim();
-      
-      let baseFontSize;
-      if (hasActualText && !panelText.fontSize) {
-        const panel = panelRects.find(p => p.panelId === panelId);
-        if (panel) {
-          baseFontSize = calculateOptimalFontSize(panelText.content, panel.width, panel.height);
-        } else {
-          baseFontSize = lastUsedTextSettings.fontSize || 26;
-        }
-      } else {
-        baseFontSize = panelText.fontSize || lastUsedTextSettings.fontSize || 26;
-      }
-      
-      return Math.round(baseFontSize * textScaleFactor);
-    }
-    
-    if (propertyName === 'strokeWidth') {
-      return currentValue || lastUsedTextSettings.strokeWidth || 2;
-    }
-    
-    if (propertyName === 'textPositionX') {
-      return currentValue !== undefined ? currentValue : (lastUsedTextSettings.textPositionX || 0);
-    }
-    
-    if (propertyName === 'textPositionY') {
-      const textPositionY = currentValue !== undefined ? currentValue : (lastUsedTextSettings.textPositionY || 0);
-      if (textPositionY <= 0) {
-        return `${Math.round(5 + (textPositionY / 100) * 5)}%`;
-      }
-      return `${Math.round(5 + (textPositionY / 100) * 95)}%`;
-    }
-    
-    if (propertyName === 'textRotation') {
-      return `${currentValue !== undefined ? currentValue : (lastUsedTextSettings.textRotation || 0)}°`;
-    }
-    
-    return currentValue || 0;
-  }, [panelTexts, lastUsedTextSettings, panelRects, calculateOptimalFontSize, textScaleFactor]);
-
-  // Slider interaction handlers
-  const handleSliderMouseDown = useCallback((panelId, propertyName) => {
-    setActiveSlider(`${panelId}-${propertyName}`);
-  }, []);
-
-  const handleSliderMouseUp = useCallback(() => {
-    setActiveSlider(null);
-  }, []);
 
   // Redraw canvas when dependencies change
   useEffect(() => {
     drawCanvas();
   }, [drawCanvas]);
 
-  // Text color scroll event handling
-  useEffect(() => {
-    const textColorElement = textColorScrollerRef.current;
-    
-    if (textColorElement) {
-      textColorElement.addEventListener('scroll', handleTextColorScroll);
-      // Check initial scroll position with a small delay to ensure content is rendered
-      setTimeout(() => {
-        handleTextColorScroll();
-      }, 100);
-    }
-    
-    const handleResize = () => {
-      handleTextColorScroll();
-    };
-    
-    window.addEventListener('resize', handleResize);
-    
-    return () => {
-      if (textColorElement) {
-        textColorElement.removeEventListener('scroll', handleTextColorScroll);
-      }
-      window.removeEventListener('resize', handleResize);
-    };
-  }, [textEditingPanel, activeTextSetting, handleTextColorScroll]);
 
-  // Check scroll position when fontStyle tab becomes active
-  useEffect(() => {
-    if (activeTextSetting === 'fontStyle' && textColorScrollerRef.current) {
-      // Multiple checks to ensure the scroll position is detected
-      setTimeout(() => handleTextColorScroll(), 50);
-      setTimeout(() => handleTextColorScroll(), 200);
-      setTimeout(() => handleTextColorScroll(), 500);
-    }
-  }, [activeTextSetting, handleTextColorScroll]);
-
-  // Effect to update localStorage when custom text color changes
-  useEffect(() => {
-    // Check if current text color is a custom color (not in the preset colors)
-    const currentTextColor = panelTexts[textEditingPanel]?.color || lastUsedTextSettings.color || '#ffffff';
-    const isCustomTextColor = !TEXT_COLOR_PRESETS.some(c => c.color === currentTextColor);
-    
-    if (isCustomTextColor && textEditingPanel) {
-      setSavedCustomTextColor(currentTextColor);
-      localStorage.setItem('memeTextCustomColor', currentTextColor);
-    }
-  }, [panelTexts, lastUsedTextSettings, textEditingPanel]);
-
-  // Focus text field when caption editor opens
-  useEffect(() => {
-    if (textEditingPanel && activeTextSetting === 'content') {
-      // Small delay to ensure the text field is rendered
-      setTimeout(() => {
-        // Get the input element directly from the inputRef
-        const inputElement = textFieldRefs.current[textEditingPanel];
-        if (inputElement) {
-          // Focus the element
-          inputElement.focus();
-          
-          // Position cursor at the end of existing text
-          const textLength = inputElement.value.length;
-          
-          // Use setTimeout to ensure focus is complete before setting selection
-          setTimeout(() => {
-            inputElement.setSelectionRange(textLength, textLength);
-            
-            // For mobile: additional steps to ensure keyboard appears
-            if ('ontouchstart' in window) {
-              inputElement.click();
-              // Double-check cursor position on mobile
-              setTimeout(() => {
-                inputElement.setSelectionRange(textLength, textLength);
-              }, 100);
-            }
-          }, 50);
-        }
-      }, 300); // Even longer delay to ensure complete rendering
-    }
-    // No cleanup needed for this effect
-  }, [textEditingPanel, activeTextSetting]);
 
   // Cleanup hover timeout on unmount
   useEffect(() => {
@@ -2995,683 +2495,19 @@ const CanvasCollagePreview = ({
             )}
             
             {/* Caption editing area - show when not in transform mode and has image, and no other panel is being edited */}
-            {!isTransformMode?.[panelId] && hasImage && (textEditingPanel === null || textEditingPanel === panelId) && (
-                <Box
-                  data-text-editor-container
-                  sx={{
-                    position: 'absolute',
-                    top: textEditingPanel === panelId ? rect.y + rect.height : rect.y + rect.height, // Always position at bottom
-                    // When collapsed: hidden, when expanded: canvas-wide positioning
-                    left: textEditingPanel === panelId ? sidePadding : rect.x + sidePadding,
-                    // When collapsed: hidden, when expanded: canvas-wide width
-                    width: textEditingPanel === panelId ? componentWidth - (sidePadding * 2) : 0,
-                    height: textEditingPanel === panelId ? 'auto' : 0,
-                    opacity: textEditingPanel === panelId ? 1 : 0,
-                    visibility: textEditingPanel === panelId ? 'visible' : 'hidden',
-                    zIndex: 20, // Above everything else including transform buttons
-                    backgroundColor: 'rgba(0, 0, 0, 0.97)',
-                    borderRadius: `${borderRadius}px`,
-                    border: '1px solid rgba(255, 255, 255, 0.3)',
-                    transition: 'all 0.3s ease-in-out', // Smooth expand/collapse
-                    overflow: 'hidden',
-                  }}
-                >
-                  {/* Expanded editing controls - only show when this panel is being edited */}
-                  {textEditingPanel === panelId && (
-                    <Box sx={{ p: 1 }}>
-                      {/* Setting tabs */}
-                      <Tabs
-                        value={activeTextSetting ? ['content', 'format', 'placement', 'fontStyle'].indexOf(activeTextSetting) : 0}
-                        onChange={(event, newValue) => {
-                          const settings = ['content', 'format', 'placement', 'fontStyle'];
-                          const newSetting = settings[newValue];
-                          setActiveTextSetting(activeTextSetting === newSetting ? null : newSetting);
-                        }}
-                        variant="fullWidth"
-                        centered
-                        sx={{
-                          mb: 2,
-                          '& .MuiTabs-indicator': {
-                            backgroundColor: '#ffffff',
-                          },
-                          '& .MuiTab-root': {
-                            color: 'rgba(255, 255, 255, 0.7)',
-                            minHeight: 44,
-                            minWidth: 0, // Allow tabs to shrink as small as needed
-                            '&.Mui-selected': {
-                              color: '#ffffff',
-                            },
-                          },
-                        }}
-                      >
-                        <Tab icon={<Edit />} />
-                        <Tab icon={<FormatSize />} />
-                        <Tab icon={<ControlCamera />} />
-                        <Tab icon={<Palette />} />
-                      </Tabs>
-
-                      {/* Content Tab */}
-                      {activeTextSetting === 'content' && (
-                        <Box sx={{ mb: 2 }}>
-                          <TextField
-                            fullWidth
-                            multiline
-                            rows={3}
-                            placeholder="Add Caption"
-                            value={panelTexts[panelId]?.content || ''}
-                            onChange={(e) => handleTextChange(panelId, 'content', e.target.value)}
-                            inputRef={(el) => {
-                              if (el) {
-                                textFieldRefs.current[panelId] = el;
-                              }
-                            }}
-                            sx={{
-                              '& .MuiInputBase-root': {
-                                backgroundColor: 'rgba(255, 255, 255, 0.9)',
-                                color: '#000000',
-                              },
-                              '& .MuiInputBase-input': {
-                                textAlign: 'center',
-                              },
-                            }}
-                          />
-                        </Box>
-                      )}
-                    
-                      {/* Format Tab */}
-                      {activeTextSetting === 'format' && (
-                        <Box sx={{ mb: 2 }}>
-                          {/* Font Size */}
-                          <Box sx={{ display: 'flex', alignItems: 'center', mb: 2 }}>
-                            <Tooltip title="Font Size" placement="left">
-                              <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'center', minWidth: 40 }}>
-                                <FormatSize sx={{ color: '#ffffff' }} />
-                              </Box>
-                            </Tooltip>
-                            <Slider
-                              value={(() => {
-                                const panelText = panelTexts[panelId] || {};
-                                const hasActualText = panelText.content && panelText.content.trim();
-                                
-                                let baseFontSize;
-                                if (hasActualText && !panelText.fontSize) {
-                                  const panel = panelRects.find(p => p.panelId === panelId);
-                                  if (panel) {
-                                    baseFontSize = calculateOptimalFontSize(panelText.content, panel.width, panel.height);
-                                  } else {
-                                    baseFontSize = lastUsedTextSettings.fontSize || 26;
-                                  }
-                                } else {
-                                  baseFontSize = panelText.fontSize || lastUsedTextSettings.fontSize || 26;
-                                }
-                                
-                                return Math.round(baseFontSize * textScaleFactor);
-                              })()}
-                              onChange={(e, value) => {
-                                if (e.type === 'mousedown') {
-                                  return;
-                                }
-                                const baseFontSize = value / textScaleFactor;
-                                handleTextChange(panelId, 'fontSize', baseFontSize);
-                              }}
-                              onMouseDown={() => handleSliderMouseDown(panelId, 'fontSize')}
-                              onMouseUp={handleSliderMouseUp}
-                              onTouchStart={() => handleSliderMouseDown(panelId, 'fontSize')}
-                              onTouchEnd={handleSliderMouseUp}
-                              min={Math.round(8 * textScaleFactor)}
-                              max={Math.round(72 * textScaleFactor)}
-                              step={1}
-                              sx={{ 
-                                flex: 1,
-                                color: '#ffffff',
-                                mx: 1,
-                              }}
-                            />
-                            <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'center', minWidth: 40 }}>
-                              {activeSlider === `${panelId}-fontSize` ? (
-                                <Typography variant="caption" sx={{ color: '#ffffff', textAlign: 'center' }}>
-                                  {getCurrentValue(panelId, 'fontSize')}
-                                </Typography>
-                              ) : (
-                                <IconButton
-                                  size="small"
-                                  onClick={() => handleResetClick('format', panelId, 'fontSize')}
-                                  disabled={isValueAtDefault(panelId, 'fontSize')}
-                                  sx={{ 
-                                    color: '#ffffff', 
-                                    p: 0.5,
-                                    '&:hover': {
-                                      backgroundColor: 'rgba(255, 255, 255, 0.1)',
-                                    },
-                                    '&.Mui-disabled': {
-                                      color: 'rgba(255, 255, 255, 0.3)',
-                                    }
-                                  }}
-                                >
-                                  <Restore fontSize="small" />
-                                </IconButton>
-                              )}
-                            </Box>
-                          </Box>
-                          
-                          {/* Stroke Width */}
-                          <Box sx={{ display: 'flex', alignItems: 'center', mb: 2 }}>
-                            <Tooltip title="Stroke Width" placement="left">
-                              <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'center', minWidth: 40 }}>
-                                <BorderOuter sx={{ color: '#ffffff' }} />
-                              </Box>
-                            </Tooltip>
-                            <Slider
-                              value={panelTexts[panelId]?.strokeWidth || lastUsedTextSettings.strokeWidth || 2}
-                              onChange={(e, value) => {
-                                if (e.type === 'mousedown') {
-                                  return;
-                                }
-                                handleTextChange(panelId, 'strokeWidth', value);
-                              }}
-                              onMouseDown={() => handleSliderMouseDown(panelId, 'strokeWidth')}
-                              onMouseUp={handleSliderMouseUp}
-                              onTouchStart={() => handleSliderMouseDown(panelId, 'strokeWidth')}
-                              onTouchEnd={handleSliderMouseUp}
-                              min={0}
-                              max={10}
-                              step={0.5}
-                              sx={{ 
-                                flex: 1,
-                                color: '#ffffff',
-                                mx: 1,
-                              }}
-                            />
-                            <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'center', minWidth: 40 }}>
-                              {activeSlider === `${panelId}-strokeWidth` ? (
-                                <Typography variant="caption" sx={{ color: '#ffffff', textAlign: 'center' }}>
-                                  {getCurrentValue(panelId, 'strokeWidth')}
-                                </Typography>
-                              ) : (
-                                <IconButton
-                                  size="small"
-                                  onClick={() => handleResetClick('format', panelId, 'strokeWidth')}
-                                  disabled={isValueAtDefault(panelId, 'strokeWidth')}
-                                  sx={{ 
-                                    color: '#ffffff', 
-                                    p: 0.5,
-                                    '&:hover': {
-                                      backgroundColor: 'rgba(255, 255, 255, 0.1)',
-                                    },
-                                    '&.Mui-disabled': {
-                                      color: 'rgba(255, 255, 255, 0.3)',
-                                    }
-                                  }}
-                                >
-                                  <Restore fontSize="small" />
-                                </IconButton>
-                              )}
-                            </Box>
-                          </Box>
-
-                        </Box>
-                      )}
-                    
-                      {/* Font Style Tab */}
-                      {activeTextSetting === 'fontStyle' && (
-                        <Box sx={{ mb: 2 }}>
-                          {/* Font Family */}
-                          <Box sx={{ display: 'flex', alignItems: 'center', mb: 2 }}>
-                            <ToggleButtonGroup
-                              value={(() => {
-                                // Normalize current weight to number for comparison
-                                const currentWeightRaw = panelTexts[panelId]?.fontWeight || lastUsedTextSettings.fontWeight || 400;
-                                let currentWeight;
-                                
-                                // Convert to number for consistent comparison
-                                if (currentWeightRaw === 'normal') {
-                                  currentWeight = 400;
-                                } else if (currentWeightRaw === 'bold') {
-                                  currentWeight = 700;
-                                } else if (typeof currentWeightRaw === 'string') {
-                                  currentWeight = parseInt(currentWeightRaw, 10) || 400;
-                                } else {
-                                  currentWeight = currentWeightRaw;
-                                }
-                                
-                                const currentStyle = panelTexts[panelId]?.fontStyle || lastUsedTextSettings.fontStyle || 'normal';
-                                const result = [];
-                                
-                                // Consider weights 500 and above as bold (more inclusive)
-                                if (currentWeight >= 500) {
-                                  result.push('bold');
-                                }
-                                
-                                if (currentStyle === 'italic') {
-                                  result.push('italic');
-                                }
-                                
-
-                                
-                                return result;
-                              })()}
-                              onChange={(event, newFormats) => {
-                                const isBold = newFormats.includes('bold');
-                                const isItalic = newFormats.includes('italic');
-                                
-                                // Update both properties in a single call to avoid timing issues
-                                const currentText = panelTexts[panelId] || {};
-                                const updatedText = {
-                                  ...currentText,
-                                  fontWeight: isBold ? 700 : 400,
-                                  fontStyle: isItalic ? 'italic' : 'normal'
-                                };
-                                
-                                if (updatePanelText) {
-                                  updatePanelText(panelId, updatedText);
-                                }
-                              }}
-                              aria-label="text formatting"
-                              sx={{ 
-                                flexShrink: 0,
-                                mr: 1,
-                                height: '42px', // Reduced height
-                                '& .MuiToggleButton-root': {
-                                  color: 'rgba(255, 255, 255, 0.7)',
-                                  borderColor: 'rgba(255, 255, 255, 0.3)',
-                                  height: '42px', // Reduced height
-                                  '&:hover': {
-                                    backgroundColor: 'rgba(255, 255, 255, 0.1)',
-                                  },
-                                  '&.Mui-selected': {
-                                    backgroundColor: 'rgba(255, 255, 255, 0.2)',
-                                    color: '#ffffff',
-                                  },
-                                }
-                              }}
-                            >
-                              <ToggleButton size='small' value="bold" aria-label="bold">
-                                <FormatBold />
-                              </ToggleButton>
-                              <ToggleButton size='small' value="italic" aria-label="italic">
-                                <FormatItalic />
-                              </ToggleButton>
-                            </ToggleButtonGroup>
-                            <FormControl sx={{ flex: 1 }}>
-                              <Select
-                                value={panelTexts[panelId]?.fontFamily || lastUsedTextSettings.fontFamily || 'Arial'}
-                                onChange={(e) => handleTextChange(panelId, 'fontFamily', e.target.value)}
-                                sx={{ 
-                                  color: '#ffffff',
-                                  fontFamily: panelTexts[panelId]?.fontFamily || lastUsedTextSettings.fontFamily || 'Arial', // Display selected font in its own font
-                                  height: '42px', // Reduced height to match toggle buttons
-                                  '& .MuiSelect-select': {
-                                    textAlign: 'left', // Left align the text
-                                    padding: '8px 14px', // Adjust padding for shorter height
-                                  },
-                                  '& .MuiOutlinedInput-notchedOutline': {
-                                    borderColor: 'rgba(255, 255, 255, 0.3)',
-                                  },
-                                  '&:hover .MuiOutlinedInput-notchedOutline': {
-                                    borderColor: 'rgba(255, 255, 255, 0.5)',
-                                  },
-                                  '&.Mui-focused .MuiOutlinedInput-notchedOutline': {
-                                    borderColor: '#ffffff',
-                                  },
-                                  '& .MuiSvgIcon-root': {
-                                    color: '#ffffff',
-                                  }
-                                }}
-                                MenuProps={{
-                                  PaperProps: {
-                                    sx: { 
-                                      maxHeight: 200,
-                                      bgcolor: 'rgba(0, 0, 0, 0.9)',
-                                      '& .MuiMenuItem-root': {
-                                        color: '#ffffff',
-                                      }
-                                    }
-                                  }
-                                }}
-                              >
-                                {fonts.map((font) => (
-                                  <MenuItem 
-                                    key={font} 
-                                    value={font}
-                                    sx={{ fontFamily: font }} // Display each font name in its own font
-                                  >
-                                    {font}
-                                  </MenuItem>
-                                ))}
-                              </Select>
-                            </FormControl>
-                          </Box>
-                          
-                          {/* Text Color */}
-                          <Box sx={{ position: 'relative' }}>
-                              <ScrollButton 
-                                direction="left" 
-                                onClick={scrollTextColorLeft}
-                                size="small"
-                                aria-label="Scroll left"
-                                sx={{ 
-                                  display: 'flex',
-                                  visibility: textColorLeftScroll ? 'visible' : 'hidden',
-                                  opacity: textColorLeftScroll ? 1 : 0,
-                                  top: '50%',
-                                  transform: 'translateY(-50%)',
-                                }}
-                              >
-                                <ChevronLeft fontSize="small" />
-                              </ScrollButton>
-                              
-                              <ScrollButton 
-                                direction="right" 
-                                onClick={scrollTextColorRight}
-                                size="small"
-                                aria-label="Scroll right"
-                                sx={{ 
-                                  display: 'flex',
-                                  visibility: textColorRightScroll ? 'visible' : 'hidden',
-                                  opacity: textColorRightScroll ? 1 : 0,
-                                  top: '50%',
-                                  transform: 'translateY(-50%)',
-                                }}
-                              >
-                                <ChevronRight fontSize="small" />
-                              </ScrollButton>
-                              <Box ref={textColorScrollerRef} sx={{ display: 'flex', alignItems: 'center', maxWidth: '100%', overflowX: 'auto', overflowY: 'hidden', position: 'relative' }}>
-                                  <Box sx={{ flex: 1, position: 'relative' }}>
-                                    <HorizontalScroller 
-                                      onScroll={handleTextColorScroll}
-                                      sx={{ 
-                                        pt: 0, 
-                                        mt: 0, 
-                                        pb: 0, 
-                                        pr: 2,
-                                        minHeight: 50,
-                                        gap: theme.spacing(1),
-                                      }}
-                                    >
-                                      {/* Custom color picker - as first option, always using saved custom color */}
-                                      <Tooltip title="Pick Custom Color" arrow>
-                                        <Box sx={{ 
-                                          position: 'relative', 
-                                          display: 'flex', 
-                                          alignItems: 'center',
-                                          height: '60%' // Match the height of the color swatches
-                                        }}>
-                                          <ColorSwatch
-                                            onClick={() => textColorPickerRef.current && textColorPickerRef.current.click()}
-                                            selected={false} // Never show this as selected
-                                            sx={{ 
-                                              position: 'relative',
-                                              flexShrink: 0, // Prevent shrinking for scrolling
-                                              // Always use the saved custom color from localStorage as background
-                                              backgroundColor: savedCustomTextColor,
-                                              // Add subtle checkerboard pattern for the color picker
-                                              backgroundImage: 'linear-gradient(45deg, rgba(200,200,200,0.2) 25%, transparent 25%), linear-gradient(-45deg, rgba(200,200,200,0.2) 25%, transparent 25%), linear-gradient(45deg, transparent 75%, rgba(200,200,200,0.2) 75%), linear-gradient(-45deg, transparent 75%, rgba(200,200,200,0.2) 75%)',
-                                              backgroundSize: '8px 8px',
-                                              backgroundPosition: '0 0, 0 4px, 4px -4px, -4px 0px',
-                                            }}
-                                          >
-                                            <Colorize fontSize="small" sx={{ color: isDarkColor(savedCustomTextColor) ? '#fff' : '#000' }} />
-                                          </ColorSwatch>
-                                          <input
-                                            type="color"
-                                            value={savedCustomTextColor}
-                                            onChange={handleCustomTextColorChange}
-                                            ref={textColorPickerRef}
-                                            style={{ 
-                                              position: 'absolute',
-                                              opacity: 0,
-                                              width: '100%',
-                                              height: '100%',
-                                              cursor: 'pointer'
-                                            }}
-                                          />
-                                        </Box>
-                                      </Tooltip>
-                                      
-                                      {/* Show the saved custom color as second option if it exists and is different from presets */}
-                                      {hasSavedCustomTextColor && (
-                                        <Tooltip title="Custom Color" arrow>
-                                          <ColorSwatch
-                                            onClick={() => handleTextChange(panelId, 'color', savedCustomTextColor)}
-                                            selected={(panelTexts[panelId]?.color || lastUsedTextSettings.color || '#ffffff') === savedCustomTextColor}
-                                            sx={{ backgroundColor: savedCustomTextColor, flexShrink: 0 }}
-                                          />
-                                        </Tooltip>
-                                      )}
-                                      
-                                      {/* Preset colors */}
-                                      {TEXT_COLOR_PRESETS.map((colorOption) => (
-                                        <Tooltip key={colorOption.color} title={colorOption.name} arrow>
-                                          <ColorSwatch
-                                            onClick={() => handleTextChange(panelId, 'color', colorOption.color)}
-                                            selected={(panelTexts[panelId]?.color || lastUsedTextSettings.color || '#ffffff') === colorOption.color}
-                                            sx={{ backgroundColor: colorOption.color, flexShrink: 0 }}
-                                          />
-                                        </Tooltip>
-                                      ))}
-                                      
-                                      {/* Spacer to ensure last items can be centered when scrolled fully */}
-                                      <Box sx={{ minWidth: 4, flexShrink: 0 }} />
-                                    </HorizontalScroller>
-                                    
-                                    {/* Visual indicators for scrolling */}
-                                    <ScrollIndicator 
-                                      direction="left" 
-                                      isVisible={textColorLeftScroll}
-                                    />
-                                    
-                                    <ScrollIndicator 
-                                      direction="right" 
-                                      isVisible={textColorRightScroll}
-                                    />
-                                  </Box>
-                              </Box>
-                          </Box>
-                        </Box>
-                      )}
-                    
-                      {/* Placement Tab */}
-                      {activeTextSetting === 'placement' && (
-                        <Box sx={{ mb: 2 }}>
-                          {/* Vertical Position */}
-                          <Box sx={{ display: 'flex', alignItems: 'center', mb: 2 }}>
-                            <Tooltip title="Vertical Position" placement="left">
-                              <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'center', minWidth: 40 }}>
-                                <SwapVert sx={{ color: '#ffffff' }} />
-                              </Box>
-                            </Tooltip>
-                            <Slider
-                              value={(() => {
-                                const textPositionY = panelTexts[panelId]?.textPositionY !== undefined ? panelTexts[panelId].textPositionY : (lastUsedTextSettings.textPositionY || 0);
-                                if (textPositionY <= 0) {
-                                  return 5 + (textPositionY / 100) * 5;
-                                }
-                                return 5 + (textPositionY / 100) * 95;
-                              })()}
-                              onChange={(e, value) => {
-                                if (e.type === 'mousedown') {
-                                  return;
-                                }
-                                let textPositionY;
-                                if (value <= 5) {
-                                  textPositionY = ((value - 5) / 5) * 100;
-                                } else {
-                                  textPositionY = ((value - 5) / 95) * 100;
-                                }
-                                handleTextChange(panelId, 'textPositionY', textPositionY);
-                              }}
-                              onMouseDown={() => handleSliderMouseDown(panelId, 'textPositionY')}
-                              onMouseUp={handleSliderMouseUp}
-                              onTouchStart={() => handleSliderMouseDown(panelId, 'textPositionY')}
-                              onTouchEnd={handleSliderMouseUp}
-                              min={0}
-                              max={100}
-                              step={1}
-                              marks={[{ value: 5 }]}
-                              sx={{ 
-                                flex: 1,
-                                color: '#ffffff',
-                                mx: 1,
-                              }}
-                            />
-                            <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'center', minWidth: 40 }}>
-                              {activeSlider === `${panelId}-textPositionY` ? (
-                                <Typography variant="caption" sx={{ color: '#ffffff', textAlign: 'center' }}>
-                                  {getCurrentValue(panelId, 'textPositionY')}
-                                </Typography>
-                              ) : (
-                                <IconButton
-                                  size="small"
-                                  onClick={() => handleResetClick('placement', panelId, 'textPositionY')}
-                                  disabled={isValueAtDefault(panelId, 'textPositionY')}
-                                  sx={{ 
-                                    color: '#ffffff', 
-                                    p: 0.5,
-                                    '&:hover': {
-                                      backgroundColor: 'rgba(255, 255, 255, 0.1)',
-                                    },
-                                    '&.Mui-disabled': {
-                                      color: 'rgba(255, 255, 255, 0.3)',
-                                    }
-                                  }}
-                                >
-                                  <Restore fontSize="small" />
-                                </IconButton>
-                              )}
-                            </Box>
-                          </Box>
-                          
-                          {/* Horizontal Position */}
-                          <Box sx={{ display: 'flex', alignItems: 'center', mb: 2 }}>
-                            <Tooltip title="Horizontal Position" placement="left">
-                              <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'center', minWidth: 40 }}>
-                                <SwapHoriz sx={{ color: '#ffffff' }} />
-                              </Box>
-                            </Tooltip>
-                            <Slider
-                              value={panelTexts[panelId]?.textPositionX !== undefined ? panelTexts[panelId].textPositionX : (lastUsedTextSettings.textPositionX || 0)}
-                              onChange={(e, value) => {
-                                if (e.type === 'mousedown') {
-                                  return;
-                                }
-                                handleTextChange(panelId, 'textPositionX', value);
-                              }}
-                              onMouseDown={() => handleSliderMouseDown(panelId, 'textPositionX')}
-                              onMouseUp={handleSliderMouseUp}
-                              onTouchStart={() => handleSliderMouseDown(panelId, 'textPositionX')}
-                              onTouchEnd={handleSliderMouseUp}
-                              min={-100}
-                              max={100}
-                              step={1}
-                              marks={[{ value: 0 }]}
-                              sx={{ 
-                                flex: 1,
-                                color: '#ffffff',
-                                mx: 1,
-                              }}
-                            />
-                            <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'center', minWidth: 40 }}>
-                              {activeSlider === `${panelId}-textPositionX` ? (
-                                <Typography variant="caption" sx={{ color: '#ffffff', textAlign: 'center' }}>
-                                  {getCurrentValue(panelId, 'textPositionX')}%
-                                </Typography>
-                              ) : (
-                                <IconButton
-                                  size="small"
-                                  onClick={() => handleResetClick('placement', panelId, 'textPositionX')}
-                                  disabled={isValueAtDefault(panelId, 'textPositionX')}
-                                  sx={{ 
-                                    color: '#ffffff', 
-                                    p: 0.5,
-                                    '&:hover': {
-                                      backgroundColor: 'rgba(255, 255, 255, 0.1)',
-                                    },
-                                    '&.Mui-disabled': {
-                                      color: 'rgba(255, 255, 255, 0.3)',
-                                    }
-                                  }}
-                                >
-                                  <Restore fontSize="small" />
-                                </IconButton>
-                              )}
-                            </Box>
-                          </Box>
-                          
-                          {/* Rotation */}
-                          <Box sx={{ display: 'flex', alignItems: 'center' }}>
-                            <Tooltip title="Rotation" placement="left">
-                              <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'center', minWidth: 40 }}>
-                                <RotateLeft sx={{ color: '#ffffff' }} />
-                              </Box>
-                            </Tooltip>
-                            <Slider
-                              value={panelTexts[panelId]?.textRotation !== undefined ? panelTexts[panelId].textRotation : (lastUsedTextSettings.textRotation || 0)}
-                              onChange={(e, value) => {
-                                if (e.type === 'mousedown') {
-                                  return;
-                                }
-                                handleTextChange(panelId, 'textRotation', value);
-                              }}
-                              onMouseDown={() => handleSliderMouseDown(panelId, 'textRotation')}
-                              onMouseUp={handleSliderMouseUp}
-                              onTouchStart={() => handleSliderMouseDown(panelId, 'textRotation')}
-                              onTouchEnd={handleSliderMouseUp}
-                              min={-180}
-                              max={180}
-                              step={1}
-                              marks={[{ value: 0 }]}
-                              sx={{ 
-                                flex: 1,
-                                color: '#ffffff',
-                                mx: 1,
-                              }}
-                            />
-                            <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'center', minWidth: 40 }}>
-                              {activeSlider === `${panelId}-textRotation` ? (
-                                <Typography variant="caption" sx={{ color: '#ffffff', textAlign: 'center' }}>
-                                  {getCurrentValue(panelId, 'textRotation')}
-                                </Typography>
-                              ) : (
-                                <IconButton
-                                  size="small"
-                                  onClick={() => handleResetClick('placement', panelId, 'textRotation')}
-                                  disabled={isValueAtDefault(panelId, 'textRotation')}
-                                  sx={{ 
-                                    color: '#ffffff', 
-                                    p: 0.5,
-                                    '&:hover': {
-                                      backgroundColor: 'rgba(255, 255, 255, 0.1)',
-                                    },
-                                    '&.Mui-disabled': {
-                                      color: 'rgba(255, 255, 255, 0.3)',
-                                    }
-                                  }}
-                                >
-                                  <Restore fontSize="small" />
-                                </IconButton>
-                              )}
-                            </Box>
-                          </Box>
-                        </Box>
-                      )}
-                    
-                      {/* Done button */}
-                      <Button
-                        fullWidth
-                        variant="contained"
-                        onClick={handleTextClose}
-                        sx={{
-                          backgroundColor: '#4CAF50',
-                          '&:hover': {
-                            backgroundColor: '#45a049',
-                          },
-                        }}
-                      >
-                        Done
-                      </Button>
-                    </Box>
-                  )}
-              </Box>
+            {!isTransformMode?.[panelId] && hasImage && textEditingPanel === panelId && (
+              <CaptionEditor
+                panelId={panelId}
+                panelTexts={panelTexts}
+                lastUsedTextSettings={lastUsedTextSettings}
+                updatePanelText={updatePanelText}
+                panelRects={panelRects}
+                calculateOptimalFontSize={calculateOptimalFontSize}
+                textScaleFactor={textScaleFactor}
+                onClose={handleTextClose}
+                rect={rect}
+                componentWidth={componentWidth}
+              />
             )}
             
 
@@ -3771,34 +2607,6 @@ const CanvasCollagePreview = ({
         />
       )}
 
-      {/* Reset Confirmation Dialog */}
-      <Dialog
-        open={resetDialogOpen}
-        onClose={handleResetCancel}
-        PaperProps={{
-          sx: {
-            backgroundColor: '#1e1e1e',
-            color: '#ffffff',
-          }
-        }}
-      >
-        <DialogTitle sx={{ color: '#ffffff' }}>
-          Reset to Default
-        </DialogTitle>
-        <DialogContent>
-          <DialogContentText sx={{ color: '#ffffff' }}>
-            Are you sure you want to reset this setting to its default value? This action cannot be undone.
-          </DialogContentText>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={handleResetCancel} sx={{ color: '#ffffff' }}>
-            Cancel
-          </Button>
-          <Button onClick={handleResetConfirm} sx={{ color: '#ffffff' }} autoFocus>
-            Reset
-          </Button>
-        </DialogActions>
-      </Dialog>
     </Box>
   );
 };
